ft_put* now return int

+++ ft_putchar.c
@@ -6,13 +6,13 @@
 /*   By: mirsella <mirsella@protonmail.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 16:39:22 by mirsella          #+#    #+#             */
-/*   Updated: 2022/11/21 16:42:07 by mirsella         ###   ########.fr       */
+/*   Updated: 2022/11/21 17:13:13 by mirsella         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "unistd.h"
 
-size_t	ft_putchar(char c)
+int	ft_putchar(char c)
 {
 	write(1, &c, 1);
 	return (1);
